# src/admin-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/projects": {
      "post": {
        "operationId": "ProjectController_create",
        "summary": "Create Project",
        "description": "This endpoint is used to create a new project with the provided details. You need to include the project details in the request body, and the API will return the created project object. \n\n**Request Body Parameters**:\n- **project_name** (required): The name of the project.\n- **start_date** (required): The start date of the project in ISO 8601 format.\n- **created_by** (required): The UUID of the user creating the project.\n- **account** (required): The ID of the account associated with the project.\n\n**How to Use**:\n- Send a POST request to the `/api/v1/admin/projects` endpoint with the project details as a JSON object in the request body.\n- Ensure all required fields are included and correctly formatted.\n\n**Response Details**:\n- On successful creation, the API returns the newly created project object with its generated ID.\n- If any required field is missing or the provided data is invalid, the API responds with a 400 Bad Request error.\n- Server errors or permission issues will result in a 500 Internal Server Error.\n\nThis endpoint allows adding new projects to the system and associating them with accounts while tracking their start dates.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "ProjectController_findAll",
        "summary": "Get Page of Project List",
        "description": "This endpoint retrieves a paginated list of all projects stored in the system. You can filter the results based on various parameters, allowing you to manage and view projects effectively.\n\n**Query Parameters**:\n- **page** (optional): The page number of the results to retrieve (default is 1).\n- **limit** (optional): The number of projects per page (default is 10).\n- **sort** (optional): The field by which to sort the projects (e.g., project_name).\n- **order** (optional): The sorting order, either ascending or descending.\n- **account** (optional): Filter projects by a specific account ID.\n\n**How to Use**:\n- Send a GET request to the `/api/v1/admin/projects` endpoint with the desired query parameters.\n- Ensure parameters are correctly formatted for filtering and sorting.\n\n**Response Details**:\n- The API returns a JSON object containing a list of projects along with pagination metadata (total count, current page, etc.).\n- If no projects match the criteria, an empty list is returned.\n\nThis endpoint is useful for retrieving and displaying all projects for better project management.",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter conditions",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/admin/projects/accounts/{account_id}/projects": {
      "get": {
        "operationId": "ProjectController_findByAccountId",
        "summary": "Get Project By Account Id",
        "description": "This endpoint retrieves a paginated list of projects associated with a specific account. You can filter and sort the projects to manage and view them effectively based on the account.\n\n**Path Parameters**:\n- **account_id** (required): The unique identifier of the account whose projects you want to retrieve.\n\n**Query Parameters**:\n- **page** (optional): The page number of the results to retrieve (default is 1).\n- **limit** (optional): The number of projects per page (default is 10).\n- **sortField** (optional): The field by which to sort the projects (e.g., project_name).\n- **sortOrder** (optional): The sorting order, either ascending or descending.\n- **filter** (optional): Filter conditions to apply when retrieving projects.\n\n**How to Use**:\n- Send a GET request to the `/api/v1/admin/projects/accounts/{account_id}/projects` endpoint, replacing `{account_id}` with the actual account ID.\n- Ensure the parameters are correctly formatted for filtering and sorting.\n\n**Response Details**:\n- The API returns a JSON object containing a list of projects associated with the specified account, along with pagination metadata (total count, current page, etc.).\n- If no projects are found for the account, an empty list is returned.\n\nThis endpoint is essential for managing projects within specific accounts, allowing for effective project tracking and oversight.",
        "parameters": [
          {
            "name": "account_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter conditions",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/admin/projects/{id}": {
      "get": {
        "operationId": "ProjectController_findOne",
        "summary": "Get Project by project id",
        "description": "This endpoint fetches the details of a specific project identified by its unique ID. It allows users to view project information in detail.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the project to retrieve.\n\n**How to Use**:\n- Send a GET request to the `/api/v1/admin/projects/{id}` endpoint, replacing `{id}` with the actual project ID.\n- Ensure the ID is correctly formatted.\n\n**Response Details**:\n- On success, the API returns a JSON object with the project’s details.\n- If the project with the specified ID is not found, a 404 Not Found error is returned.\n\nThis endpoint is essential for accessing detailed information about specific projects.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "ProjectController_update",
        "summary": "Update Project",
        "description": "This endpoint updates the details of an existing project identified by its unique ID. You need to provide the updated project details in the request body.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the project to update.\n\n**Request Body Parameters**:\n- **project_name** (optional): The updated name of the project.\n- **start_date** (optional): The updated start date of the project in ISO 8601 format.\n- **account** (optional): The updated account ID associated with the project.\n\n**How to Use**:\n- Send a PUT request to the `/api/v1/admin/projects/{id}` endpoint with the updated project details as a JSON object in the request body.\n- Ensure that the project ID is in the URL and the request body includes valid fields.\n\n**Response Details**:\n- On successful update, the API returns the updated project object.\n- If the project is not found, a 404 Not Found error is returned.\n- Validation errors will result in a 400 Bad Request error.\n\nThis endpoint is crucial for maintaining accurate and up-to-date project information.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "operationId": "ProjectController_delete",
        "summary": "Delete Project",
        "description": "This endpoint deletes an existing project from the system using its unique ID. It is a destructive operation that permanently removes the project.\n\n**Path Parameters**:\n- **id** (required): The unique identifier of the project to delete.\n\n**How to Use**:\n- Send a DELETE request to the `/api/v1/admin/projects/{id}` endpoint, replacing `{id}` with the actual project ID.\n- Ensure that you have the necessary permissions to perform this operation.\n\n**Response Details**:\n- On successful deletion, the API returns a confirmation message indicating the project has been deleted.\n- If the project is not found, a 404 Not Found error is returned.\n- Permission errors will result in a 403 Forbidden error.\n\nThis endpoint is essential for managing project lifecycles by allowing deletions of obsolete or incorrect projects.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Admin Service",
    "description": "Admin Service - admin services handle account, project, user and role management",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "account_id": {
            "type": "string",
            "description": "Account Id for which project is to be created"
          },
          "project_name": {
            "type": "string",
            "description": "Name of project to be created"
          },
          "start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Start date of the project"
          }
        },
        "required": [
          "account_id",
          "project_name",
          "start_date"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "account_id": {
            "type": "string",
            "description": "Account Id for which project is to be created"
          },
          "project_name": {
            "type": "string",
            "description": "Name of project to be created"
          },
          "start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Start date of the project"
          }
        }
      },
...
    }
  }
}